+++ internal/bundler/bundler_packagejson_test.go
@@ -1194,3 +1194,189 @@ func TestPackageJsonNeutralExplicitMainFields(t *testing.T) {
 		},
 	})
 }
+
+func TestPackageJsonExportsErrorInvalidModuleSpecifier(t *testing.T) {
+	packagejson_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import 'pkg1'
+				import 'pkg2'
+				import 'pkg3'
+				import 'pkg4'
+				import 'pkg5'
+				import 'pkg6'
+			`,
+			"/Users/user/project/node_modules/pkg1/package.json": `
+				{ "exports": { ".": "./%%" } }
+			`,
+			"/Users/user/project/node_modules/pkg2/package.json": `
+				{ "exports": { ".": "./%2f" } }
+			`,
+			"/Users/user/project/node_modules/pkg3/package.json": `
+				{ "exports": { ".": "./%2F" } }
+			`,
+			"/Users/user/project/node_modules/pkg4/package.json": `
+				{ "exports": { ".": "./%5c" } }
+			`,
+			"/Users/user/project/node_modules/pkg5/package.json": `
+				{ "exports": { ".": "./%5C" } }
+			`,
+			"/Users/user/project/node_modules/pkg6/package.json": `
+				{ "exports": { ".": "./%31.js" } }
+			`,
+			"/Users/user/project/node_modules/pkg6/1.js": `
+				console.log(1)
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/src/entry.js: error: Could not resolve "pkg1" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg1/package.json: note: The module specifier "./%%" is invalid
+Users/user/project/src/entry.js: error: Could not resolve "pkg2" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg2/package.json: note: The module specifier "./%2f" is invalid
+Users/user/project/src/entry.js: error: Could not resolve "pkg3" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg3/package.json: note: The module specifier "./%2F" is invalid
+Users/user/project/src/entry.js: error: Could not resolve "pkg4" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg4/package.json: note: The module specifier "./%5c" is invalid
+Users/user/project/src/entry.js: error: Could not resolve "pkg5" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg5/package.json: note: The module specifier "./%5C" is invalid
+`,
+	})
+}
+
+func TestPackageJsonExportsErrorInvalidPackageConfiguration(t *testing.T) {
+	packagejson_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import 'pkg1'
+				import 'pkg2/foo'
+			`,
+			"/Users/user/project/node_modules/pkg1/package.json": `
+				{ "exports": { ".": false } }
+			`,
+			"/Users/user/project/node_modules/pkg2/package.json": `
+				{ "exports": { "./foo": false } }
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/node_modules/pkg1/package.json: warning: This value must be a string, an object, an array, or null
+Users/user/project/node_modules/pkg2/package.json: warning: This value must be a string, an object, an array, or null
+Users/user/project/src/entry.js: error: Could not resolve "pkg1" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg1/package.json: note: The package configuration has an invalid value here
+Users/user/project/src/entry.js: error: Could not resolve "pkg2/foo" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg2/package.json: note: The package configuration has an invalid value here
+`,
+	})
+}
+
+func TestPackageJsonExportsErrorInvalidPackageTarget(t *testing.T) {
+	packagejson_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import 'pkg1'
+				import 'pkg2'
+				import 'pkg3'
+			`,
+			"/Users/user/project/node_modules/pkg1/package.json": `
+				{ "exports": { ".": "invalid" } }
+			`,
+			"/Users/user/project/node_modules/pkg2/package.json": `
+				{ "exports": { ".": "../pkg3" } }
+			`,
+			"/Users/user/project/node_modules/pkg3/package.json": `
+				{ "exports": { ".": "./node_modules/pkg" } }
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/src/entry.js: error: Could not resolve "pkg1" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg1/package.json: note: The package target "invalid" is invalid
+Users/user/project/src/entry.js: error: Could not resolve "pkg2" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg2/package.json: note: The package target "../pkg3" is invalid
+Users/user/project/src/entry.js: error: Could not resolve "pkg3" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg3/package.json: note: The package target "./node_modules/pkg" is invalid
+`,
+	})
+}
+
+func TestPackageJsonExportsErrorPackagePathNotExported(t *testing.T) {
+	packagejson_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import 'pkg1/foo'
+			`,
+			"/Users/user/project/node_modules/pkg1/package.json": `
+				{ "exports": { ".": {} } }
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/src/entry.js: error: Could not resolve "pkg1/foo" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg1/package.json: note: The path "./foo" is not exported by "pkg1"
+`,
+	})
+}
+
+func TestPackageJsonExportsErrorModuleNotFound(t *testing.T) {
+	packagejson_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import 'pkg1'
+			`,
+			"/Users/user/project/node_modules/pkg1/package.json": `
+				{ "exports": { ".": "./foo.js" } }
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/src/entry.js: error: Could not resolve "pkg1" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg1/package.json: note: The module "./foo.js" was not found
+`,
+	})
+}
+
+func TestPackageJsonExportsErrorUnsupportedDirectoryImport(t *testing.T) {
+	packagejson_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import 'pkg1'
+				import 'pkg2'
+			`,
+			"/Users/user/project/node_modules/pkg1/package.json": `
+				{ "exports": { ".": "./foo/" } }
+			`,
+			"/Users/user/project/node_modules/pkg2/package.json": `
+				{ "exports": { ".": "./foo" } }
+			`,
+			"/Users/user/project/node_modules/pkg2/foo/bar.js": `
+				console.log(bar)
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/src/entry.js: error: Could not resolve "pkg1" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg1/package.json: note: The module "./foo" was not found
+Users/user/project/src/entry.js: error: Could not resolve "pkg2" (mark it as external to exclude it from the bundle)
+Users/user/project/node_modules/pkg2/package.json: note: Importing the directory "./foo" is not supported
+`,
+	})
+}